refactor: reorganize test structure and update file paths for integration tests
